+++ src/wican_cli/commands/__init__.py
@@ -3,6 +3,7 @@
 from wican_cli.commands.autopid import register as register_autopid
 from wican_cli.commands.config import register as register_config
 from wican_cli.commands.logs import register as register_logs
+from wican_cli.commands.profile import register as register_profile
 from wican_cli.commands.protocol import register as register_protocol
 from wican_cli.commands.reboot import register as register_reboot
 from wican_cli.commands.sleep import register as register_sleep
@@ -16,4 +17,5 @@ ALL_REGISTERS = [
     register_logs,
     register_protocol,
     register_autopid,
+    register_profile,
 ]